# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/deployments/jobs": {
      "get": {
        "tags": [
          "runtime"
        ],
        "operationId": "list_deployment_jobs",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeploymentJob"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "runtime"
        ],
        "operationId": "create_deployment_job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeploymentJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentJob"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/deployments/jobs/{job_id}": {
      "get": {
        "tags": [
          "runtime"
        ],
        "operationId": "get_deployment_job",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentJob"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/deployments/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "runtime"
        ],
        "operationId": "cancel_deployment_job",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentJob"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/leaders/{leader_agent_id}/executors": {
      "get": {
        "tags": [
          "leaders"
        ],
        "operationId": "list_leader_executors",
        "parameters": [
          {
            "name": "leader_agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeaderExecutor"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "leaders"
        ],
        "operationId": "update_leader_executors",
        "parameters": [
          {
            "name": "leader_agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLeaderExecutorsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeaderExecutor"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/settings/auth": {
      "get": {
        "tags": [
          "settings"
        ],
        "operationId": "get_auth_settings",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSettings"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "operationId": "update_auth_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSettings"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/integrations": {
      "get": {
        "tags": [
          "settings"
        ],
        "operationId": "get_integration_settings",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationSettings"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "operationId": "update_integration_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrationSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationSettings"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/ports": {
      "get": {
        "tags": [
          "settings"
        ],
        "operationId": "get_port_settings",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortSettings"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "operationId": "update_port_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortSettings"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/runtime": {
      "get": {
        "tags": [
          "settings"
        ],
        "operationId": "get_runtime_settings",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeSettings"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "operationId": "update_runtime_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeSettings"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/{user_id}/role": {
      "patch": {
        "tags": [
          "auth"
        ],
        "operationId": "update_user_role",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentKind": {
        "type": "string",
        "enum": [
          "hermes",
          "java_agent"
        ]
      },
...
      "AgentRole": {
        "type": "string",
        "enum": [
          "developer",
          "tester",
          "it_lead",
          "custom"
        ]
      },
...
      "AuthSettings": {
        "type": "object",
        "required": [
          "access_token_ttl_minutes",
          "refresh_token_ttl_days",
          "refresh_cookie_name",
          "refresh_cookie_secure",
          "refresh_cookie_same_site",
          "refresh_cookie_path"
        ],
        "properties": {
          "access_token_ttl_minutes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "refresh_cookie_domain": {
            "type": [
              "string",
              "null"
            ]
          },
          "refresh_cookie_name": {
            "type": "string"
          },
          "refresh_cookie_path": {
            "type": "string"
          },
          "refresh_cookie_same_site": {
            "type": "string"
          },
          "refresh_cookie_secure": {
            "type": "boolean"
          },
          "refresh_token_ttl_days": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "CreateDeploymentJobRequest": {
        "type": "object",
        "required": [
          "job_kind",
          "title"
        ],
        "properties": {
          "agent_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "detail": {},
          "job_kind": {
            "$ref": "#/components/schemas/DeploymentJobKind"
          },
          "runtime_kind": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AgentKind"
              }
            ]
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "DeploymentJob": {
        "type": "object",
        "required": [
          "id",
          "job_kind",
          "state",
          "title",
          "detail",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "agent_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "created_at": {
            "$ref": "#/components/schemas/String"
          },
          "detail": {},
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "job_kind": {
            "$ref": "#/components/schemas/DeploymentJobKind"
          },
          "last_error": {
            "type": [
              "string",
              "null"
            ]
          },
          "requested_by_user_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "runtime_kind": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AgentKind"
              }
            ]
          },
          "state": {
            "$ref": "#/components/schemas/DeploymentJobState"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "$ref": "#/components/schemas/String"
          }
        }
      },
      "DeploymentJobKind": {
        "type": "string",
        "enum": [
          "provision",
          "runtime_update"
        ]
      },
      "DeploymentJobState": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed",
          "cancelled"
        ]
      },
...
      "IntegrationSettings": {
        "type": "object",
        "required": [
          "project_workflow_status"
        ],
        "properties": {
          "github_remote": {
            "type": [
              "string",
              "null"
            ]
          },
          "project_workflow_status": {
            "type": "string"
          },
          "project_workflow_url": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "LeaderExecutor": {
        "type": "object",
        "required": [
          "leader_agent_id",
          "executor_agent_id",
          "executor_name",
          "executor_display_name",
          "executor_profile",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "$ref": "#/components/schemas/String"
          },
          "created_by_user_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "executor_agent_id": {
            "type": "string",
            "format": "uuid"
          },
          "executor_display_name": {
            "type": "string"
          },
          "executor_name": {
            "type": "string"
          },
          "executor_profile": {
            "$ref": "#/components/schemas/AgentRole"
          },
          "leader_agent_id": {
            "type": "string",
            "format": "uuid"
          },
          "namespace_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "workflow_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "PortSettings": {
        "type": "object",
        "required": [
          "backend_port",
          "frontend_port",
          "agent_port_base",
          "agent_port_stride"
        ],
        "properties": {
          "agent_port_base": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "agent_port_stride": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "backend_port": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "frontend_port": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "RuntimeSettings": {
        "type": "object",
        "required": [
          "agents_root",
          "hermes_source",
          "hermes_command",
          "java_agent_source",
          "java_agent_command"
        ],
        "properties": {
          "agents_root": {
            "type": "string"
          },
          "hermes_command": {
            "type": "string"
          },
          "hermes_source": {
            "type": "string"
          },
          "java_agent_command": {
            "type": "string"
          },
          "java_agent_source": {
            "type": "string"
          }
        }
      },
...
      "String": {
        "type": "string"
      },
      "SystemRole": {
        "type": "string",
        "enum": [
          "admin",
          "operator",
          "user"
        ]
      },
...
      "UpdateLeaderExecutorsRequest": {
        "type": "object",
        "required": [
          "executor_ids"
        ],
        "properties": {
          "executor_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "UpdateUserRoleRequest": {
        "type": "object",
        "required": [
          "role"
        ],
        "properties": {
          "role": {
            "$ref": "#/components/schemas/SystemRole"
          }
        }
      },
...
      "UserResponse": {
        "type": "object",
        "required": [
          "id",
          "email",
          "username",
          "display_name",
          "system_role",
          "is_system_admin",
          "is_active"
        ],
        "properties": {
          "display_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_active": {
            "type": "boolean"
          },
          "is_system_admin": {
            "type": "boolean"
          },
          "system_role": {
            "$ref": "#/components/schemas/SystemRole"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}